- Move beta features out of unnecessary beta namespaces/folders
